Implemented Dao Services the Correct Way
