+++ app/src/main/java/com/stvayush/keepitclean/framework/datasource/database/NoteDao.kt
@@ -1,0 +1,20 @@
+package com.stvayush.keepitclean.framework.datasource.database
+
+
+/** Just a placeholder for now, will configure it while implementing/setting Up room */
+
+const val NOTE_ORDER_ASC: String = ""
+const val NOTE_ORDER_DESC: String = "-"
+const val NOTE_FILTER_TITLE = "title"
+const val NOTE_FILTER_DATE_CREATED = "created_at"
+
+const val ORDER_BY_ASC_DATE_UPDATED = NOTE_ORDER_ASC + NOTE_FILTER_DATE_CREATED
+const val ORDER_BY_DESC_DATE_UPDATED = NOTE_ORDER_DESC + NOTE_FILTER_DATE_CREATED
+const val ORDER_BY_ASC_TITLE = NOTE_ORDER_ASC + NOTE_FILTER_TITLE
+const val ORDER_BY_DESC_TITLE = NOTE_ORDER_DESC + NOTE_FILTER_TITLE
+
+const val NOTE_PAGINATION_PAGE_SIZE = 30
+
+interface NoteDao{
+
+}
